Merge remote-tracking branch 'origin/user/herault/homogeneous_precision'
into candidate-master.

The remote branch includes:

* the new ODE system for floating objects;
* the new homogeneous precision code.

Also fix a minor compile isse on Linux (-ode instead of -lode in the
Makefile).

The code has only been compile-tested so far.

Conflicts:
	Makefile
	src/Cube.cc
	src/DamBreak3D.cc
	src/DamBreak3D.h
	src/EnergyGenerator.h
	src/GPUSph.cc
	src/Object.h
	src/OdeObjects.h
	src/OpenChannel.cc
	src/Options.h
	src/ParticleSystem.cc
	src/ParticleSystem.h
	src/Problem.cc
	src/Problem.h
	src/Test.h
	src/TestTopo.cc
	src/TopoCube.h
	src/VTKWriter.cc
	src/buildneibs.cu
	src/buildneibs.cuh
	src/buildneibs_kernel.cu
	src/euler.cu
	src/euler.cuh
	src/euler_kernel.def
	src/forces.cu
	src/forces.cuh
	src/forces_kernel.cu
	src/forces_kernel.visc.inc
	src/particledefine.h

+++ src/CustomTextWriter.cc
@@ -53,7 +53,7 @@ CustomTextWriter::~CustomTextWriter()
     }
 }
 
-void CustomTextWriter::write(uint numParts, const float4 *pos, const float4 *vel,
+void CustomTextWriter::write(uint numParts, const double4 *pos, const float4 *vel,
 					const particleinfo *info, const float3 *vort, float t,const bool testpoints, const float4 *normals)
 {
 	string filename, full_filename;
