+++ src/Options.h
@@ -32,13 +32,13 @@
 using namespace std;
 
 struct Options {
-	string problem; // problem name
-	int device;  // which device to use
-	string dem; // DEM file to use
-	string dir; // directory where data will be saved
-	bool console; // run in console (no GUI)
-	float deltap; // deltap
-	float tend; // simulation end
+	string	problem; // problem name
+	int		device;  // which device to use
+	string	dem; // DEM file to use
+	string	dir; // directory where data will be saved
+	bool	console; // run in console (no GUI)
+	double	deltap; // deltap
+	float	tend; // simulation end
 	Options(void) :
 		problem(),
 		device(-1),
