+++ src/TextWriter.cc
@@ -54,7 +54,7 @@ TextWriter::~TextWriter()
     }
 }
 
-void TextWriter::write(uint numParts, const float4 *pos, const float4 *vel,
+void TextWriter::write(uint numParts, const double4 *pos, const float4 *vel,
 					const particleinfo *info, const float3 *vort, float t, const bool testpoints, const float4 *normals)
 {
 	string filename, full_filename;
